重构上传下载速度统计 (#823)
* Remove CacheItemView and MediaCachePresentation

This is a leftover from the cache management rewrite

* 重构上传和下载速度统计 (torrent, media cache)

* Refactor: Optimize FileSize class

This commit optimizes the FileSize class by:

- Forbid direct constructor calls
- Simplifying calculations by using Float instead of Double for intermediate values.
- Handling Unspecified values as 0 in operations.
- Adding unary plus and minus operators.
- Improving the toString() method to produce more concise output.

* 修复创建缓存后直接显示完成, 实际上并没有完成

* Avoid potential division by zero in FileStats

* 修复缓存管理 UI 未同步

* fix 持久

+++ data-sources/bt/dmhy/src/impl/protocol/ListParser.kt
@@ -106,10 +106,10 @@ internal object ListParser {
         private fun createFileSize(size: String): FileSize {
             // 430.1 MB
             return when {
-                size.endsWith("GB") -> size.removeSuffix("GB").toDouble().gigaBytes
-                size.endsWith("MB") -> size.removeSuffix("MB").toDouble().megaBytes
-                size.endsWith("KB") -> size.removeSuffix("KB").toDouble().kiloBytes
-                size.endsWith("B") -> size.removeSuffix("B").toDouble().bytes
+                size.endsWith("GB") -> size.removeSuffix("GB").toFloat().gigaBytes
+                size.endsWith("MB") -> size.removeSuffix("MB").toFloat().megaBytes
+                size.endsWith("KB") -> size.removeSuffix("KB").toFloat().kiloBytes
+                size.endsWith("B") -> size.removeSuffix("B").toFloat().bytes
                 else -> throw IllegalArgumentException("Unrecognized size pattern: $size")
             }
         }
